vonmises: disable numba
Doesn't seem to help

+++ tests/test_vonmises.py
@@ -110,7 +110,7 @@ class TestBvMKernelDensityEstimator(object):
 
     def test_estimate(self):
         kde = BvMKernelDensityEstimator(
-            n_bins=256, k1=10, k2=20, k3=-2, dtype=np.float32, batchsize=64
+            n_bins=256, k1=10, k2=20, k3=-2, dtype=np.float64, batchsize=64
         )
         phi_psi = np.random.uniform(-np.pi, np.pi, size=(500, 2))
         p = kde(phi_psi[:, 0], phi_psi[:, 1])
